adi/adrv9002.py: Remove mcs function
The attribute "multi_chip_sync" will have to be set in a separate
thread for each device.
With the newer Linux drivers, setting the "multi_chip_sync" attribute,
will hold the iio_attr setting process until the device exits the MCS
state, or timeout occurs.
This is why it is needed to start each attr setting in a separate thread
and also check if the thread is alive, after the MCS process should be
complete to validate that the MCS was completed successfully.

Signed-off-by: AndreiGrozav <andrei.grozav@analog.com>