jupiter_sync.py: Update data capture
The updated system functionality will automatically change the sync
module functionality from MCS to data transfer sync. So, no need for
direct reg writes at the sync address.

Removed the loopback functionality which was used for debug purposes.

The data capture must be done on separate threads. This reasoning is that
after we arm the data paths the DMA's will not be able to transfer data
until a trigger event occurs and that in it's current form the capture
data mechanism waits until a buffer is captured.
So, in the new format we will, arm the ADC/DAC, request  a transfer, send
a trigger signal in the form of a sync request.

To ensure the sync between channels we used the MIMO mode(2RX2TX), more
exactly the RX channels will use only one DMA pack cores, rather than,
each channel with it's on pack and DMA.

Signed-off-by: AndreiGrozav <andrei.grozav@analog.com>